requirements.txt
streamlit==1.42.1
pandas==2.2.3
pillow==11.1.0
openpyxl==3.1.5